protection/http: finer-grained instrumentation
Mainly for debugging purposes, split the existing file in two so that we can
have one that is not instrumented thanks to the `sqreen:ignore` directive.
